LSDP/pyproject.toml
[project]
name = "exercises"
version = "1.0.0"
description = "Large Scale Drone Perception course"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "codetiming>=1.3.0",
    "exifread>=3.0.0",
    "icecream>=2.0.0",
    "matplotlib>=3.10.0",
    "opencv-python>=4.11.0",
    "pandas>=2.3.3",
    "plotly>=5.0.0",
    "pyopengl>=3.1.5",
    "scikit-learn>=1.6.0",
    "youtube-dl>=2021.12.17",
    "yt-dlp>=2026.3.17",
]
